Fix build on 64-bit Haiku
- Fix path to look for libnetwork to work also on non-hybrid systems
- Remove references to /boot/common as it is not used anymore
- Fix a few type errors
- Use compiler provided va_copy when available

+++ beos/window.cpp
@@ -1275,7 +1275,7 @@ void gui_get_clipboard(char **buffer, size_t *length)
 		clip = be_clipboard->Data();
 		if (clip) {
 			const char *text;
-			int32 textlen;
+			ssize_t textlen;
 			if (clip->FindData("text/plain", B_MIME_TYPE, 
 				(const void **)&text, &textlen) >= B_OK) {
 				*buffer = (char *)malloc(textlen);
